requirements.txt
Django>=4.2,<5.0
django-environ>=0.10.0
django-crispy-forms>=2.0
pillow>=10.0.0
whitenoise>=6.5.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.7
django-widget-tweaks>=1.4.12
django-allauth>=0.54.0
python-dotenv>=1.0.0